fix: Troca o tipo float por Decimal na classe do pedidos (views) #46

+++ app/views/order.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from django.conf import settings
 from django.utils import timezone
 from django.utils.dateparse import parse_date
@@ -327,8 +329,8 @@ def EvaluateOrder(request, pk):
             )
 
             product.status = form_status
-            available_quantity = float(form_available_quantity)
-            product_balance = float(product.call_product.balance)
+            available_quantity = Decimal(form_available_quantity)
+            product_balance = product.call_product.balance
 
             if available_quantity > product_balance:
                 messages.error(
